+++ code/svb-docker/python/svb_web/core/urls.py
@@ -4,8 +4,8 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-     path('', views.public.views.index, name="index"),
-    path('account/', views.admin.articles.account, name="account"),
-    path('news_editor/', views.admin.articles.news_editor),
-    path('article/<str:article_headline>/', views.public.articles.article, name='article'),
+     path('', views.public.common.index, name="index"),
+    path('account/', views.public.customer.account, name="account"),
+    path('news_editor/', views.internal.author.news_editor),
+    path('article/<str:article_headline>/', views.public.visitor.article, name='article'),
 ]
