+++ code/svb-docker/python/svb_web/core/views/__init__.py
@@ -1,2 +1,7 @@
-from .public.views import *
-from .admin.articles import *
+from .public.common import *
+from .public.customer import *
+from .public.visitor import *
+
+from .internal.author import *
+from .internal.banker import *
+from .internal.display import *
