+++ code/svb-docker/python/svb_web/core/views/internal/author.py
@@ -48,13 +48,3 @@ def news_editor(request):
         'article_form': article_form,
     }
     return render(request, 'news_editor.html', context=context)
-
-
-def index(request):
-    context = {}
-    return render(request, "index.html", context)
-
-
-def account(request, account_number=None):
-    context = {}
-    return render(request, "account.html", context)
