+++ code/svb-docker/python/svb_web/core/views/public/common.py
@@ -1,0 +1,6 @@
+from django.shortcuts import render
+
+# Create your views here.
+def index(request):
+    context = {}
+    return render(request, "index.html", context)
