+++ code/svb-docker/python/svb_web/core/views/public/customer.py
@@ -1,0 +1,6 @@
+from django.shortcuts import render
+
+# Create your views here.
+def account(request, account_number=None):
+    context = {}
+    return render(request, "account.html", context)
